requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pdfminer.six==20221105

# Google Cloud
google-cloud-storage==2.13.0
google-cloud-functions==1.13.3
google-cloud-scheduler==2.13.0
google-cloud-logging==3.8.0

# AI and ML
google-cloud-aiplatform==1.36.4
google-cloud-vision==3.4.5
google-cloud-texttospeech==2.14.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Video Generation
moviepy==1.0.3
Pillow==10.1.0
numpy==1.26.1
matplotlib==3.8.1
pysrt==1.1.2  # For subtitle handling
ffmpeg-python==0.2.0  # For advanced video processing

# Runway API (for Gen-2 video generation)
runway-python==0.7.0  # Unofficial Runway API client
tenacity==8.2.3  # For retry logic

# Web and API
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1